Создание пакаджей, перемещение файлов по ним. Начало архитектуры.
