+++ app/src/androidTest/java/ru/pseudonimb/filmsearch/MainActivityTest.kt
@@ -16,6 +16,8 @@ import org.hamcrest.Matchers.allOf
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import ru.pseudonimb.filmsearch.view.MainActivity
+import ru.pseudonimb.filmsearch.view.rv_viewholders.FilmViewHolder
 
 @RunWith(AndroidJUnit4::class)
 class MainActivityTest {
