+++ app/src/main/java/ru/pseudonimb/filmsearch/view/fragments/DetailsFragment.kt
@@ -1,4 +1,4 @@
-package ru.pseudonimb.filmsearch
+package ru.pseudonimb.filmsearch.view.fragments
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,7 +6,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import ru.pseudonimb.filmsearch.R
 import ru.pseudonimb.filmsearch.databinding.FragmentDetailsBinding
+import ru.pseudonimb.filmsearch.domain.Film
 
 class DetailsFragment : Fragment() {
     private lateinit var film: Film
