+++ app/src/main/java/ru/pseudonimb/filmsearch/view/fragments/FavoritesFragment.kt
@@ -1,4 +1,4 @@
-package ru.pseudonimb.filmsearch
+package ru.pseudonimb.filmsearch.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,7 +6,12 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import ru.pseudonimb.filmsearch.view.rv_adapters.FilmListRecyclerAdapter
+import ru.pseudonimb.filmsearch.view.rv_adapters.TopSpacingItemDecoration
 import ru.pseudonimb.filmsearch.databinding.FragmentFavoritesBinding
+import ru.pseudonimb.filmsearch.domain.Film
+import ru.pseudonimb.filmsearch.utils.AnimationHelper
+import ru.pseudonimb.filmsearch.view.MainActivity
 
 class FavoritesFragment : Fragment() {
     private lateinit var filmsAdapter: FilmListRecyclerAdapter
@@ -32,7 +37,7 @@ class FavoritesFragment : Fragment() {
         AnimationHelper.performFragmentCircularRevealAnimation(binding.favoritesFragmentRoot, requireActivity(), 1)
 
         binding.favoritesRecycler.apply {
-            filmsAdapter = FilmListRecyclerAdapter(object : FilmListRecyclerAdapter.OnItemClickListener{
+            filmsAdapter = FilmListRecyclerAdapter(object : FilmListRecyclerAdapter.OnItemClickListener {
                 override fun click(film: Film) {
                     (requireActivity() as MainActivity).launchDetailsFragment(film)
                 }
