+++ app/src/main/java/ru/pseudonimb/filmsearch/view/fragments/SelectionsFragment.kt
@@ -1,15 +1,16 @@
-package ru.pseudonimb.filmsearch
+package ru.pseudonimb.filmsearch.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import ru.pseudonimb.filmsearch.databinding.FragmentWatchLaterBinding
+import ru.pseudonimb.filmsearch.databinding.FragmentSelectionsBinding
+import ru.pseudonimb.filmsearch.utils.AnimationHelper
 
-class WatchLaterFragment : Fragment() {
+class SelectionsFragment : Fragment() {
 
-    private var _binding: FragmentWatchLaterBinding? = null
+    private var _binding: FragmentSelectionsBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -17,15 +18,14 @@ class WatchLaterFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentWatchLaterBinding.inflate(inflater, container, false)
+        _binding = FragmentSelectionsBinding.inflate(inflater, container, false)
         val view = binding.root
         return view
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        AnimationHelper.performFragmentCircularRevealAnimation(binding.watchLaterFragmentRoot, requireActivity(), 3)
+        AnimationHelper.performFragmentCircularRevealAnimation(binding.selectionsFragmentRoot, requireActivity(), 4)
     }
 
     override fun onDestroyView() {
