+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:background="@drawable/background_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
     <FrameLayout
         android:id="@+id/fragment_placeholder"
