+++ app/src/main/res/layout/fragment_watch_later.xml
@@ -7,6 +7,6 @@
     android:visibility="invisible"
     tools:visibility="visible"
     android:background="@drawable/background_watch_later"
-    tools:context=".WatchLaterFragment">
+    tools:context=".view.fragments.WatchLaterFragment">
 
 </FrameLayout>
